feat: implementacion de alta y cancelacion de reservas

+++ src/main/java/unla/tp/tp_distribuidos/repositories/IVehiculoRepository.java
@@ -8,7 +8,7 @@ import unla.tp.tp_distribuidos.models.Vehiculo;
 
 @Repository("VehiculoRepository")
 public interface IVehiculoRepository extends JpaRepository<Vehiculo, Integer>{
-    Optional<Vehiculo> findById(int id);
+    Optional<Vehiculo> findById(Long id);
     Optional<Vehiculo> findByPatente(String patente);
     List<Vehiculo> findAll();
 }
